# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Motorcycle Rental & Catalog API 🏍️",
    "version": "1.0.0",
    "description": "API REST robusta en Node.js y TypeScript para la gestión de catálogo de motocicletas, reservas con validación de fechas, opiniones de usuarios y autenticación basada en JWT."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desarrollo Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Ingresa tu token JWT en formato: `Bearer <token>`"
      }
    },
    "schemas": {
...
      "Motorcycle": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "brand": { "type": "string", "example": "Yamaha" },
          "model": { "type": "string", "example": "MT-07" },
          "year": { "type": "integer", "example": 2024 },
          "price": { "type": "number", "format": "float", "description": "Precio diario de alquiler", "example": 95.0 },
          "engineCapacity": { "type": "integer", "description": "Cilindrada en cc", "example": 689 },
          "imageUrl": { "type": "string", "format": "uri", "nullable": true, "example": "https://example.com/yamaha.jpg" },
          "description": { "type": "string", "nullable": true, "example": "Una moto naked de peso medio." },
          "available": { "type": "boolean", "default": true, "example": true }
        }
      },
...
      "Review": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "userId": { "type": "integer", "example": 2 },
          "motorcycleId": { "type": "integer", "example": 1 },
          "rating": { "type": "integer", "minimum": 1, "maximum": 5, "example": 5 },
          "comment": { "type": "string", "nullable": true, "example": "Excelente manejo y torque." }
        }
      }
    }
  },
  "paths": {
...
    "/api/motorcycles": {
      "get": {
        "tags": ["Motocicletas"],
        "summary": "Listar motocicletas con filtros de búsqueda",
        "parameters": [
          { "name": "brand", "in": "query", "description": "Filtrar por marca", "required": false, "schema": { "type": "string" } },
          { "name": "year", "in": "query", "description": "Filtrar por año", "required": false, "schema": { "type": "integer" } },
          { "name": "available", "in": "query", "description": "Filtrar por disponibilidad ('true' o 'false')", "required": false, "schema": { "type": "string" } },
          { "name": "minPrice", "in": "query", "description": "Precio diario mínimo", "required": false, "schema": { "type": "number" } },
          { "name": "maxPrice", "in": "query", "description": "Precio diario máximo", "required": false, "schema": { "type": "number" } },
          { "name": "search", "in": "query", "description": "Buscar en marca, modelo o descripción", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Catálogo de motocicletas obtenido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Motorcycle" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Motocicletas"],
        "summary": "Añadir una motocicleta al catálogo",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["brand", "model", "year", "price", "engineCapacity"],
                "properties": {
                  "brand": { "type": "string", "example": "Honda" },
                  "model": { "type": "string", "example": "CBR 650R" },
                  "year": { "type": "integer", "example": 2023 },
                  "price": { "type": "number", "example": 110.0 },
                  "engineCapacity": { "type": "integer", "example": 649 },
                  "imageUrl": { "type": "string", "example": "https://example.com/honda.jpg" },
                  "description": { "type": "string", "example": "Moto deportiva ágil y potente." },
                  "available": { "type": "boolean", "default": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Motocicleta creada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Motorcycle added successfully" },
                    "motorcycle": { "$ref": "#/components/schemas/Motorcycle" }
                  }
                }
              }
            }
          },
          "400": { "description": "Campos requeridos faltantes" },
          "403": { "description": "Permisos de Administrador requeridos" }
        }
      }
    },
    "/api/motorcycles/{id}": {
      "get": {
        "tags": ["Motocicletas"],
        "summary": "Obtener detalles de una motocicleta específica por ID (incluye opiniones)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Detalles de la motocicleta",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Motorcycle" },
                    {
                      "type": "object",
                      "properties": {
                        "reviews": {
                          "type": "array",
                          "items": {
                            "allOf": [
                              { "$ref": "#/components/schemas/Review" },
                              {
                                "type": "object",
                                "properties": {
                                  "user": {
                                    "type": "object",
                                    "properties": {
                                      "id": { "type": "integer" },
                                      "name": { "type": "string" }
                                    }
                                  }
                                }
                              }
                            ]
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": { "description": "Motocicleta no encontrada" }
        }
      },
      "put": {
        "tags": ["Motocicletas"],
        "summary": "Actualizar detalles de una motocicleta",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "brand": { "type": "string" },
                  "model": { "type": "string" },
                  "year": { "type": "integer" },
                  "price": { "type": "number" },
                  "engineCapacity": { "type": "integer" },
                  "imageUrl": { "type": "string" },
                  "description": { "type": "string" },
                  "available": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Motocicleta actualizada exitosamente" },
          "404": { "description": "Motocicleta no encontrada" },
          "403": { "description": "Permisos de Administrador requeridos" }
        }
      },
      "delete": {
        "tags": ["Motocicletas"],
        "summary": "Eliminar una motocicleta del catálogo",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Motocicleta eliminada del catálogo" },
          "404": { "description": "Motocicleta no encontrada" },
          "403": { "description": "Permisos de Administrador requeridos" }
        }
      }
    },
...
  }
}